How to kill package kit
PackageKit is the most annoying part of openSUSE initially (or at least it was for me). It gets in the way of any kind of update or package manager update or app install. It's just real annoying. These are the two commands I found that can get rid of it temporarily. I'll also be adding this to the beginning of every other script that might encounter issues with PackageKit